pyproject.toml
[project]
name = "west-env"
version = "0.1.0"
description = "Workspace-first west extension for reproducible Zephyr build environments with optional container-backed execution"
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0",
    "west>=1.5.0",
]

[project.entry-points."west.commands"]
env = "west_commands.env:EnvCommand"

[project.optional-dependencies]
test = ["pytest>=7.0"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
  "west_env",
  "west_commands",
]
